+++ src/main.rs
@@ -31,7 +31,7 @@ use std::{
 use indicatif::{MultiProgress, ProgressBar, ProgressStyle};
 use console::Term;
 
-use crate::common::{LabelRecog, AssemblyCode};
+use crate::common::{LabelRecog, AssemblyCode, errors::ExitErrorCode};
 
 fn cli_interface() -> ArgMatches {
     #[allow(non_upper_case_globals)]
@@ -141,7 +141,7 @@ fn main() {
     let matches = cli_interface();
 
     let vals: Vec<&PathBuf> = matches.get_many::<PathBuf>("input")
-    .expect("At least one assembly input file is required")
+    .unwrap()
     .collect();
 
     let mut parsed_vector: Vec<AssemblyCode<LabelRecog>> = vec![];
@@ -173,7 +173,7 @@ fn main() {
             Ok(val) => string_vector.push(val),
             Err(msg) => {
                 error!("Could not read '{}': {}", file.as_path().to_string_lossy(), msg);
-                std::process::exit(1)
+                std::process::exit(66)
             },
         };
     }
@@ -189,7 +189,7 @@ fn main() {
             Ok(val) => parsed_vector.push(val.1),
             Err(e) => {
                 error!("{e}");
-                std::process::exit(1)
+                std::process::exit(65)
             },
         }
     }
@@ -209,7 +209,7 @@ fn main() {
         Ok(linked) => linked,
         Err(e) => {
             error!("{e}");
-            std::process::exit(1)
+            std::process::exit(e.get_err_code())
         },
     };
 
@@ -222,7 +222,7 @@ fn main() {
         Ok(instr_list) => instr_list,
         Err(e) => {
             error!("{e}");
-            std::process::exit(1)
+            std::process::exit(e.get_err_code())
         }
     };
 
@@ -238,7 +238,7 @@ fn main() {
 
     if let Err(e) = translator::translate_and_present(outpath, translatable_code, comment, outfmt, (*depth, width)) {
         error!("{e}");
-        std::process::exit(1)
+        std::process::exit(e.get_err_code())
     };
 
     let line = if outfmt != "debug" {