+++ src/parser/text.rs
@@ -21,7 +21,7 @@ use nom::{
 use std::cmp::Ordering;
 use std::collections::BTreeMap;
 use std::borrow::Cow;
-use log::{debug, error};
+use log::{debug, error, log_enabled};
 
 use crate::parser::{
     handle_label_defs, instructions::parse_instruction, literals::{
@@ -374,6 +374,19 @@ fn handle_multiline_immediate<'a>(
     false
 }
 
+#[inline]
+fn debug_operation_vec(macro_in: &MacroInstr, pointer: &usize, mid_list: &[Operation]) {
+    let mut print_string = format!("Expanded '{macro_in}' at {} into '[{}", *pointer, mid_list[0]);
+
+    for op in &mid_list[1..] {
+        print_string.push_str(&("; ".to_owned() + &op.to_string()));
+    }
+
+    print_string.push_str("]'");
+
+    debug!("{print_string}");
+}
+
 fn translate_macros<'a>(
     macro_in: &MacroInstr,
     instr_list: &mut Vec<Operation<'a>>,
@@ -394,7 +407,9 @@ fn translate_macros<'a>(
                 mid_list.push(Instruction::Addi(reg1.to_owned(), Reg::G10, 0).into());
             }
 
-            debug!("Expanded '{:?}' at {} into '{:?}'", macro_in, *pointer, mid_list);
+            if log_enabled!(log::Level::Debug) {
+                debug_operation_vec(macro_in, pointer, &mid_list);
+            }
 
             incorporate_changes(instr_list, &mut mid_list, &mut right_list, accumulator, pointer, label);
         },
@@ -410,7 +425,9 @@ fn translate_macros<'a>(
                 mid_list.push(Instruction::Addi(reg1.to_owned(), Reg::G10, 0).into());
             }
 
-            debug!("Expanded '{:?}' at {} into '{:?}'", macro_in, *pointer, mid_list);
+            if log_enabled!(log::Level::Debug) {
+                debug_operation_vec(macro_in, pointer, &mid_list);
+            }
 
             incorporate_changes(instr_list, &mut mid_list, &mut right_list, accumulator, pointer, label);
         },
@@ -419,7 +436,7 @@ fn translate_macros<'a>(
             let mut imm_used = *imm;
 
             if handle_multiline_immediate(&mut imm_used, label.clone(), pointer, instr_list, &Instruction::Addi(reg.to_owned(), Reg::G0, *imm)) {
-                debug!("Expanded '{:?}' at {} into '{:?}'", macro_in, *pointer - 1, instr_list.last());
+                debug!("Expanded '{macro_in}' at {} into '{}'", *pointer - 1, instr_list.last().unwrap());
                 return
             }
 
@@ -432,7 +449,7 @@ fn translate_macros<'a>(
             instr_list.insert(*pointer,
             Instruction::Addi(reg.to_owned(), reg.to_owned(), imm_used).into());
 
-            debug!("Expanded '{:?}' at {} into '[{:?}, {:?}]'", macro_in, *pointer - 1, instr_list[*pointer-1], instr_list[*pointer]);
+            debug!("Expanded '{macro_in}' at {} into '[{}; {}]'", *pointer - 1, instr_list[*pointer-1], instr_list[*pointer]);
 
             *pointer += 1;
             *accumulator += 1;
@@ -458,7 +475,7 @@ fn translate_macros<'a>(
             let mut imm_used = *imm;
 
             if handle_multiline_immediate(&mut imm_used, label.clone(), pointer, instr_list, &Instruction::Addi(reg.to_owned(), Reg::G0, *imm)) {
-                debug!("Expanded '{:?}' at {} into '{:?}'", macro_in, *pointer - 1, instr_list.last());
+                debug!("Expanded '{macro_in}' at {} into '{}'", *pointer - 1, instr_list.last().unwrap());
                 return
             }
 
@@ -471,7 +488,7 @@ fn translate_macros<'a>(
             instr_list.insert(*pointer,
             Instruction::Addi(reg.to_owned(), reg.to_owned(), imm_used).into());
 
-            debug!("Expanded '{:?}' at {} into '[{:?}, {:?}]'", macro_in, *pointer - 1, instr_list[*pointer-1], instr_list[*pointer]);
+            debug!("Expanded '{macro_in}' at {} into '[{}; {}]'", *pointer - 1, instr_list[*pointer-1], instr_list[*pointer]);
 
             *pointer += 1;
             *accumulator += 1;
@@ -487,7 +504,7 @@ fn translate_macros<'a>(
             instr_list.insert(*pointer,
             MacroInstr::Addi(reg.to_owned(), reg.to_owned(), targ_labl.to_string(), Part::Lower).into());
 
-            debug!("Expanded '{:?}' at {} into '[{:?}, {:?}]'", macro_in, *pointer - 1, instr_list[*pointer-1], instr_list[*pointer]);
+            debug!("Expanded '{macro_in}' at {} into '[{}; {}]'", *pointer - 1, instr_list[*pointer-1], instr_list[*pointer]);
 
             *pointer += 1;
             *accumulator += 1;
@@ -497,7 +514,7 @@ fn translate_macros<'a>(
             let mut imm_used = *imm;
 
             if handle_multiline_immediate(&mut imm_used, label.clone(), pointer, instr_list, &Instruction::Jalr(Reg::G1, Reg::G0, *imm)) {
-                debug!("Expanded '{:?}' at {} into '{:?}'", macro_in, *pointer - 1, instr_list.last());
+                debug!("Expanded '{macro_in}' at {} into '{}'", *pointer - 1, instr_list.last().unwrap());
                 return
             }
 
@@ -510,7 +527,7 @@ fn translate_macros<'a>(
             instr_list.insert(*pointer,
             Instruction::Jalr(Reg::G1, Reg::G1, imm_used).into());
 
-            debug!("Expanded '{:?}' at {} into '[{:?}, {:?}]'", macro_in, *pointer - 1, instr_list[*pointer-1], instr_list[*pointer]);
+            debug!("Expanded '{macro_in}' at {} into '[{}; {}]'", *pointer - 1, instr_list[*pointer-1], instr_list[*pointer]);
 
             *pointer += 1;
             *accumulator += 1;
@@ -520,7 +537,7 @@ fn translate_macros<'a>(
             let mut imm_used = *imm;
 
             if handle_multiline_immediate(&mut imm_used, label.clone(), pointer, instr_list, &Instruction::Jalr(Reg::G0, Reg::G0, *imm)) {
-                debug!("Expanded '{:?}' at {} into '{:?}'", macro_in, *pointer - 1, instr_list.last());
+                debug!("Expanded '{macro_in}' at {} into '{}'", *pointer - 1, instr_list.last().unwrap());
                 return
             }
 
@@ -533,7 +550,7 @@ fn translate_macros<'a>(
             instr_list.insert(*pointer,
             Instruction::Jalr(Reg::G0, Reg::G6, imm_used).into());
 
-            debug!("Expanded '{:?}' at {} into '[{:?}, {:?}]'", macro_in, *pointer - 1, instr_list[*pointer-1], instr_list[*pointer]);
+            debug!("Expanded '{macro_in}' at {} into '[{}; {}]'", *pointer - 1, instr_list[*pointer-1], instr_list[*pointer]);
 
             *pointer += 1;
             *accumulator += 1;
@@ -549,7 +566,7 @@ fn translate_macros<'a>(
             instr_list.insert(*pointer,
             MacroInstr::Jalr(Reg::G1, Reg::G1, targ_labl.to_string(), Part::Lower).into());
 
-            debug!("Expanded '{:?}' at {} into '[{:?}, {:?}]'", macro_in, *pointer - 1, instr_list[*pointer-1], instr_list[*pointer]);
+            debug!("Expanded '{macro_in}' at {} into '[{}; {}]'", *pointer - 1, instr_list[*pointer-1], instr_list[*pointer]);
 
             *pointer += 1;
             *accumulator += 1;
@@ -565,7 +582,7 @@ fn translate_macros<'a>(
             instr_list.insert(*pointer,
             MacroInstr::Jalr(Reg::G0, Reg::G6, targ_labl.to_string(), Part::Lower).into());
 
-            debug!("Expanded '{:?}' at {} into '[{:?}, {:?}]'", macro_in, *pointer - 1, instr_list[*pointer-1], instr_list[*pointer]);
+            debug!("Expanded '{macro_in}' at {} into '[{}; {}]'", *pointer - 1, instr_list[*pointer-1], instr_list[*pointer]);
 
             *pointer += 1;
             *accumulator += 1;
@@ -585,7 +602,9 @@ fn translate_macros<'a>(
                 acc -= 4;
             }
 
-            debug!("Expanded '{:?}' at {} into '{:?}'", macro_in, *pointer, mid_list);
+            if log_enabled!(log::Level::Debug) {
+                debug_operation_vec(macro_in, pointer, &mid_list);
+            }
 
             *accumulator += (mid_list.len() - 1) as i128;
             *pointer += mid_list.len();
@@ -613,7 +632,9 @@ fn translate_macros<'a>(
 
             mid_list.push(Instruction::Addi(Reg::G2, Reg::G2, regs.len() as i32 * 4).into());
 
-            debug!("Expanded '{:?}' at {} into '{:?}'", macro_in, *pointer, mid_list);
+            if log_enabled!(log::Level::Debug) {
+                debug_operation_vec(macro_in, pointer, &mid_list);
+            }
 
             *accumulator += (mid_list.len() - 1) as i128;
             *pointer += mid_list.len();
@@ -636,7 +657,9 @@ fn translate_macros<'a>(
                 mid_list.push(instrni.clone().into());
             }
 
-            debug!("Expanded '{:?}' at {} into '{:?}'", macro_in, *pointer, mid_list);
+            if log_enabled!(log::Level::Debug) {
+                debug_operation_vec(macro_in, pointer, &mid_list);
+            }
 
             *accumulator += (*num - 1) as i128;
             *pointer += *num as usize;
@@ -669,7 +692,9 @@ fn translate_macros<'a>(
                 }
             }
 
-            debug!("Expanded '{:?}' at {} into '{:?}'", macro_in, *pointer, mid_list);
+            if log_enabled!(log::Level::Debug) {
+                debug_operation_vec(macro_in, pointer, &mid_list);
+            }
 
             *accumulator += (space_needed - 1) as i128;
             *pointer += space_needed;
@@ -745,7 +770,7 @@ pub fn parse<'a>(input: &'a str, subroutines: &mut Option<&mut Subroutines>, sym
 
         match &mut parsed {
             (Some(label), Some(instr)) => {
-                debug!("({instr_counter}) - Parsed label '{label}' and instruction '{:?}'", instr);
+                debug!("({instr_counter}) - Parsed label '{label}' and instruction '{instr}'");
                 if let Err(e) = handle_label_defs(label, symbol_map, LabelType::Address, instr_counter) {
                     error!("{e}");
                     std::process::exit(1)
@@ -829,7 +854,7 @@ pub fn parse<'a>(input: &'a str, subroutines: &mut Option<&mut Subroutines>, sym
                 instr_list.push(instr.to_owned());
             },
             (None, Some(instr)) => {
-                debug!("({instr_counter}) - Parsed instruction '{:?}'", instr);
+                debug!("({instr_counter}) - Parsed instruction '{instr}'");
                 match instr {
                     Operation::Macro(macro_in) => handle_label_refs(macro_in, subroutines, symbol_map),
                     Operation::Instr(instr_in) => {
@@ -915,7 +940,7 @@ pub fn parse<'a>(input: &'a str, subroutines: &mut Option<&mut Subroutines>, sym
                 instr_counter += 1;
                 instr_list.push(Operation::Labl(std::borrow::Cow::Borrowed(label)));
             },
-            (None, None) => debug!("{instr_counter} - Parsed nothing!"),
+            (None, None) => debug!("({instr_counter}) - Parsed nothing!"),
         }
 
         if rest.trim().is_empty() {
@@ -942,8 +967,6 @@ pub fn parse<'a>(input: &'a str, subroutines: &mut Option<&mut Subroutines>, sym
     }
 
     expand_instrs(symbol_map, &mut instr_list);
-
-    debug!("Finished parser step");
 
     Ok(("", instr_list))
 }